// projects/codx-od/src/lib/oddashboard/models/OD_DispatchDashBoard.ts
export class OD_DispatchDashBoard {
  dispatchNo: string;
  dispatchOn: Date;
  dispatchType: string;
  refDate: Date;
  createdOn: Date;
  refNo: string;
  dispatchTypeName: string;
  urgency: string;
  urgencyName: string;
  security: string;
  securityName: string;
  status: string;
  statusName: string;
  approveStatus: string;
  approveStatusName: string;
  sendMode: string;
  sendModeName: string;
  agencyID: string;
  agencyName: string;
  deadline?: Date;
  dueDate?: Date;
  createdBy: string;
  createdByName: string;
  positionID: string;
  positionName: string;
  orgUnitID: string;
  orgUnitName: string;
  divisionID: string;
  divisionName: string;
  departmentID: string;
  departmentName: string;
  companyID: string;
  companyName: string;
  owner: string;
  ownerName: string;
  oPositionID: string;
  oPositionName: string;
  oOrgUnitID: string;
  oOrgUnitName: string;
  oDivisionID: string;
  oDivisionName: string;
  oDepartmentID: string;
  oDepartmentName: string;
  oCompanyID: string;
  oCompanyName: string;
}